requirements.txt
python-telegram-bot==13.7
requests
pytz
datetime
urllib3
python-dotenv
